grpc-pb: Move BitSet to utils
